+++ packages/@wesjet/client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @wesjet/client
 
+## 0.0.17
+
+### Patch Changes
+
+- Wesjet Api
+- Updated dependencies
+  - @wesjet/core@0.0.17
+
 ## 0.0.9
 
 ### Patch Changes