+++ packages/@wesjet/sanity-store/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @wesjet/sanity-store
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @wesjet/utils@0.0.17
+  - @wesjet/core@0.0.17
+
 ## 0.0.9
 
 ### Patch Changes