+++ packages/@wesjet/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wesjet/utils
 
+## 0.0.17
+
+### Patch Changes
+
+- Wesjet Api
+
 ## 0.0.9
 
 ### Patch Changes