+++ packages/wesjet/CHANGELOG.md
@@ -1,5 +1,17 @@
 # wesjet
 
+## 0.0.17
+
+### Patch Changes
+
+- Wesjet Api
+- Updated dependencies
+  - @wesjet/client@0.0.17
+  - @wesjet/maker@0.0.17
+  - @wesjet/utils@0.0.17
+  - @wesjet/core@0.0.17
+  - @wesjet/cli@0.0.17
+
 ## 0.0.9
 
 ### Patch Changes